move : styles 폴더 common으로 이동
