+++ pages/_app.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import { Global, ThemeProvider } from '@emotion/react';
 import type { AppProps } from 'next/app';
-import GlobalStyles from '../src/styles/globalStyles';
-import useTheme from '../src/styles/hooks/useTheme';
-import mode from '../src/styles/theme';
+import GlobalStyles from '../src/common/styles/globalStyles';
+import useTheme from '../src/common/styles/hooks/useTheme';
+import mode from '../src/common/styles/theme';
 
 export default function App({ Component, pageProps }: AppProps) {
   const [theme ] = useTheme();
