+++ src/common/styles/globalStyles.ts
