+++ src/common/styles/hooks/useTheme.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useState } from 'react';
-import useClientLayoutEffect from '../../common/hooks/useClientLayoutEffect';
+import useClientLayoutEffect from '../../hooks/useClientLayoutEffect';
 import { ModeType, UseTheme } from '../types/useTheme.types';
 
 const useTheme : ()=>UseTheme = ()=>{
